backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai==1.0.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
google-cloud-secret-manager==2.20.0
firebase-admin==6.5.0
python-dotenv==1.0.1
websockets>=13.0
pydantic==2.9.0
